refactored pandas
