+++ raphtory/src/python/graph/mod.rs
@@ -2,7 +2,7 @@ pub mod algorithm_result;
 pub mod edge;
 pub mod graph;
 pub mod graph_with_deletions;
-pub mod pandas;
 pub mod properties;
 pub mod vertex;
 pub mod views;
+pub mod pandas;
